refactor: rename syntux mod to parse
