+++ src/parse/session.rs
@@ -222,7 +222,7 @@ impl ParseSess {
     }
 }
 
-// Methods that should be restricted within the syntux module.
+// Methods that should be restricted within the parse module.
 impl ParseSess {
     pub(super) fn emit_diagnostics(&self, diagnostics: Vec<Diagnostic>) {
         for diagnostic in diagnostics {
