+++ src/rewrite.rs
@@ -7,9 +7,9 @@ use rustc_ast::ptr;
 use rustc_span::Span;
 
 use crate::config::{Config, IndentStyle};
+use crate::parse::session::ParseSess;
 use crate::shape::Shape;
 use crate::skip::SkipContext;
-use crate::syntux::session::ParseSess;
 use crate::visitor::SnippetProvider;
 use crate::FormatReport;
